Fix `descr` for nested sequence
All `descr` for structured arrays must be lists, but if they contain a
nested structured array the `descr` must be a tuple. This means that
when we extract the `descr` for a nested sequence it should be wrapped
in a list if we want to treat the array as independent.