Fix sequence unpacking
When a sequence is unpacked the data is now returned inside an
`IterData` object. Also, since we now iterate over a template of the
object, we can check for sequences by looking at the type instead of
peeking for bytes in the DODS response. This way, the `peek` method was
removed.