Removed slicing of sequence children
This patch removes the ability of slicing a children from a sequence.
I don't think slicing a deep object makes sense: a children `foo` might
have values `[0, 1]` in one record and `[]` in the next. How would
indexing work in this case?

Now slicing can be applied only to the outermost sequence.