Run tests from `pydap.tests` instead of `pydap/src`
I changed the default behavior in nose so that the tests are run from
the `pydap.tests` module, instead of running from the directory
`pydap/src`. This allows testing 3rd party packages that also have tests
in the shared namespace.

I'm not sure if this is a good thing, since the test suite now is not
restricted to the module. The good side is that it allows us to test all
handlers/responses from a single location, which is good IMHO. On the
other hand, unit tests may fail even if pydap is correct.