Changed the way `close` works
I've redesigned the way the `close` method works. The idea behind the
method is that handlers sometimes will want to close a dataset after the
request is completed. For example, the NetCDF handler will open the
file, leaving it open until the request/response cycle is complete. In
order to do this, the handler defines a `close` method that closes the
file pointer.

To call the method, the base handler will assign the method it to the
response application created. Webob will then take care of calling
`close` after the request/response cycle.