Ignore int16 tests
Eric Firing has asked for int16 DAP variables to be represented as int16
arrays in Numpy, even though they are actually transferred as int32.
This makes sense, but I'm not happy with the code that I wrote to handle
this. In the meantime, I'm leaving the corresponding unit tests here for
the future.